% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_convert_carsimr.R
\name{cpp_convert_carsimr}
\alias{cpp_convert_carsimr}
\title{Convert a matrix or list object to carsimr class}
\usage{
cpp_convert_carsimr(object)
}
\arguments{
\item{object}{A matrix or list object representing a carsimr simulation.}
}
\value{
An object with carsimr class, either a carsimr_list or carsimr matrix.
}
\description{
This function takes an object and converts it to the carsimr class. If the
object is a list, each element is converted to the carsimr class, and the
entire list is returned as a carsimr_list. If the object is a matrix, it is
returned as a single carsimr matrix. If the input is not a list or matrix,
an error is thrown.
}
