% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_grid.R
\name{initialize_grid}
\alias{initialize_grid}
\title{Create the initialize grid function to return class instance}
\usage{
initialize_grid(rho, dims, prob_blue)
}
\arguments{
\item{rho}{\itemize{
\item The density of cars in the grid
}}

\item{dims}{\itemize{
\item A vector of length 2 representing the dimensions of the grid.
}}

\item{prob_blue}{\itemize{
\item The proportion of blue squares in the grid,
a number between 0 and 1.
}}
}
\value{
This returns a matrix with blue, red, and no cars represented
by 0,1,2 respectively.
}
\description{
Create the initialize grid function to return class instance
}
