% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialize_grid_cpp}
\alias{initialize_grid_cpp}
\title{Create a matrix of values 0, 1, and 2 in C++}
\usage{
initialize_grid_cpp(rho, dims, prob_blue)
}
\arguments{
\item{rho}{\itemize{
\item The density of cars in the grid
}}

\item{dims}{\itemize{
\item A vector of length 2 representing the dimensions of the grid.
}}

\item{prob_blue}{\itemize{
\item The proportion, between 0 and 1
of blue squares out of total number of cars.
}}
}
\value{
A NumericMatrix representing the grid,
where blue and red cars are represented by 1 and 2,
respectively, and no cars are represented by 0.
}
\description{
This function generates a grid of the specified dimensions and
density with randomly distributed red and blue cars.
}
\examples{
initialize_grid_cpp(rho = 56, c(10, 10), 0.4)
initialize_grid_cpp(rho = 0.50, c(3, 3), 0.96)
}
