% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_blue.R
\name{move_blue}
\alias{move_blue}
\title{Move blue cars upward in a grid}
\usage{
move_blue(t_grid)
}
\arguments{
\item{t_grid}{A matrix representing the initial configuration
of the cars.}
}
\value{
A matrix representing the configuration of the
cars after moving the blue cars upward.
}
\description{
This function moves the blue cars upward in a grid.
It takes a matrix as input and returns a new matrix
with the blue cars moved one cell up if possible.
If a blue car is in the top row, it is moved to the
bottom row if the cell in the same column in the bottom row
is empty. Red cars are not moved.
}
