% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_blue.R
\name{move_blue_car}
\alias{move_blue_car}
\title{Move a single blue car upward in a grid}
\usage{
move_blue_car(i, j, t_grid, new_grid, dims)
}
\arguments{
\item{i}{The row index of the car.}

\item{j}{The column index of the car.}

\item{t_grid}{The original grid.}

\item{new_grid}{The new grid being constructed.}

\item{dims}{The dimensions of the grid.}
}
\value{
The updated new grid.
}
\description{
This function moves a single blue car upward in a grid.
If the car is in the top row, it is moved to the bottom row
if the cell in the same column in the bottom row is empty.
If the car can move one cell up, it is moved to that cell.
Otherwise, it stays in the same position.
}
