% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_cars.R
\name{move_cars}
\alias{move_cars}
\title{Move cars function}
\usage{
move_cars(t_grid, trials)
}
\arguments{
\item{t_grid}{A matrix representing the starting grid
configuration with 0 for empty cells, 1 for blue cars, and 2 for red cars.}

\item{trials}{An integer specifying the number of trials to move the cars.}
}
\value{
A list of matrices representing the grid configuration
after each trial, including the initial grid.
}
\description{
This function moves the red and blue cars in a grid for
a specified number of trials.
}
