% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{move_cars_cpp}
\alias{move_cars_cpp}
\title{Move both cars multiple trials in C++}
\usage{
move_cars_cpp(grid, trials)
}
\arguments{
\item{grid}{A NumericMatrix representing the starting grid
with 0 for empty cells, 1 for blue cars, and 2 for red cars.}

\item{trials}{An integer specifying the number of trials to move the cars.}
}
\value{
A List of NumericMatrices representing the grid configuration
after each trial, including the initial grid.
}
\description{
Moves the red and blue cars in a grid for
a specified number of trials in C++.
}
