% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_red.R
\name{move_red}
\alias{move_red}
\title{Move the red cars horizontally rightward}
\usage{
move_red(t_grid)
}
\arguments{
\item{t_grid}{The traffic simulation grid to be updated.}
}
\value{
A new grid with the updated positions of the cars.
}
\description{
This function moves the red cars in a traffic simulation grid
horizontally rightward by one unit if there is no car in the way.
Blue cars are not moved. The function returns a new grid with the
updated positions of the cars.
}
