% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_red.R
\name{move_red_car}
\alias{move_red_car}
\title{Move a single red car horizontally rightward in a grid}
\usage{
move_red_car(i, j, t_grid, new_grid, dims)
}
\arguments{
\item{i}{The row index of the car.}

\item{j}{The column index of the car.}

\item{t_grid}{The original grid.}

\item{new_grid}{The new grid being constructed.}

\item{dims}{The dimensions of the grid.}
}
\value{
The updated new grid.
}
\description{
This function moves a single red car horizontally rightward in a grid.
If the car is in the rightmost column, it is moved to the leftmost column
if the cell in the same row in the leftmost column is empty.
If the car can move one cell rightward, it is moved to that cell.
Otherwise, it stays in the same position.
}
