% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.carsimr_list}
\alias{plot.carsimr_list}
\title{Plot List of carsimr output}
\usage{
\method{plot}{carsimr_list}(x, y, pause, ...)
}
\arguments{
\item{x}{A list of matrices containing the carsimr outputs}

\item{y}{Ignored}

\item{pause}{The pause time (in seconds) between each
plot image print}

\item{...}{(optional) Additional arguments for the plot.carsimr function}
}
\value{
A sequence of plots protraying the carsimr outputs
}
\description{
Plots a list of carsimr outputs, each containing
an image with blue and red representing the cars
}
\examples{
grid <- initialize_grid(0.50, c(3, 3), 0.5)
grid_list <- move_cars(grid, 3)
plot(grid_list, pause = 1)
}
