% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_grid.R
\name{initialize_grid}
\alias{initialize_grid}
\title{Create a matrix of values 0, 1, and 2}
\usage{
initialize_grid(rho, dims, prob_blue)
}
\arguments{
\item{rho}{\itemize{
\item The density of cars in the grid
}}

\item{dims}{\itemize{
\item A vector of length 2 representing the dimensions of the grid.
}}

\item{prob_blue}{\itemize{
\item The proportion, between 0 and 1
of blue squares out of total number of cars.
}}
}
\value{
Returns a matrix representing the grid,
where blue and red cars are represented by 1 and 2,
respectively, and no cars are represented by 0.
}
\description{
Create a matrix of values 0, 1, and 2
}
\examples{
initialize_grid(rho = 56, c(10, 10), 0.4)
initialize_grid(rho = 0.50, c(3, 3), 0.96)
}
